+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Add new method :meth:`pycov.CovFrame.chr_prefix`.
 * Add new property ``contigs`` to :class:`pybed.BedFrame`.
 * Add new method :meth:`pybed.BedFrame.chr_prefix`.
+* Add new methods :meth:`pybed.BedFrame.copy_meta` and :meth:`pybed.BedFrame.sort`.
 
 0.22.0 (2021-09-04)
 -------------------